third_party/blink/web_tests/external/wpt/fetch/private-network-access/service-worker-update.https.window-expected.txt

This is a testharness.js-based test.
PASS update public to local: failed preflight.
FAIL update public to local: success. assert_equals: update error expected (undefined) undefined but got (string) "TypeError"
PASS update public to private: failed preflight.
FAIL update public to private: success. assert_equals: update error expected (undefined) undefined but got (string) "TypeError"
PASS update public to public: success.
Harness: the test ran to completion.
